@@ -18,7 +18,7 @@
         "phpdocumentor/reflection-docblock": "^5.2",
         "ramsey/uuid": "^4.1",
         "sensio/framework-extra-bundle": "^5.1",
-        "symfony/asset": "4.4.*",
+        "symfony/asset": "5.3.*",
         "symfony/console": "4.4.*",
         "symfony/dotenv": "5.3.*",
         "symfony/expression-language": "4.4.*",
