@@ -4,6 +4,7 @@
 namespace App\Model\User\Entity\User;
 
 
+use DateTimeImmutable;
 use Doctrine\Common\Collections\ArrayCollection;
 use Doctrine\ORM\Mapping as ORM;
 
@@ -26,55 +27,55 @@ class User
      * @ORM\Column(type="user_user_id")
      * @ORM\Id
      */
-    private $id;
+    private Id $id;
 
     /**
-     * @var \DateTimeImmutable
+     * @var DateTimeImmutable
      * @ORM\Column(type="datetime_immutable")
      */
-    private $date;
+    private DateTimeImmutable $date;
     /**
      * @var Email|null
      * @ORM\Column(type="user_user_email", nullable=true)
      */
-    private $email;
+    private ?Email $email;
 
     /**
      * @var string|null
      * @ORM\Column(type="string", nullable=true, name="password_hash")
      */
-    private $passwordHash;
+    private ?string $passwordHash;
 
     /**
      * @var string|null
      * @ORM\Column(type="string", nullable=true, name="confirm_token")
      */
-    private $confirmToken;
+    private ?string $confirmToken;
 
     /**
      * @var ResetToken|null
      * @ORM\Embedded(class="ResetToken", columnPrefix="reset_token_")
      */
-    private $resetToken;
+    private ?ResetToken $resetToken;
 
     /**
      * @var string|null
      * @ORM\Column(type="string", length=12)
      */
-    private $status;
+    private ?string $status;
 
     /**
      * @var Role
      * @ORM\Column(type="user_user_role")
      */
-    private $role;
+    private Role $role;
     /**
      * @var Network[]|ArrayCollection
      * @ORM\OneToMany(targetEntity="Network", mappedBy="user", orphanRemoval=true , cascade={"persist"})
      */
     private $networks;
 
-    public function __construct(Id $id, \DateTimeImmutable $date)
+    public function __construct(Id $id, DateTimeImmutable $date)
     {
         $this->id = $id;
         $this->date = $date;
@@ -83,7 +84,7 @@ class User
         $this->networks = new ArrayCollection();
     }
 
-    public static function signUpByEmail(Id $id, \DateTimeImmutable  $date, Email $email, string $hash, string $token): self
+    public static function signUpByEmail(Id $id, DateTimeImmutable  $date, Email $email, string $hash, string $token): self
     {
         $user = new self($id, $date);
         $user->email = $email;
@@ -110,7 +111,7 @@ class User
         $this->networks->add(new Network($this, $network, $identity));
     }
 
-    public function requestPasswordReset(ResetToken $token, \DateTimeImmutable $date): void
+    public function requestPasswordReset(ResetToken $token, DateTimeImmutable $date): void
     {
         if(!$this->isActive()){
             throw new \DomainException('User is not active');
@@ -125,7 +126,7 @@ class User
         $this->resetToken = $token;
     }
 
-    public function passwordReset(\DateTimeImmutable $date, string $hash): void
+    public function passwordReset(DateTimeImmutable $date, string $hash): void
     {
         if(!$this->resetToken){
             throw new \DomainException('Resetting is not requested!');
@@ -174,7 +175,7 @@ class User
         return $this->id;
     }
 
-    public function getDate(): \DateTimeImmutable
+    public function getDate(): DateTimeImmutable
     {
         return $this->date;
     }
