@@ -1,18 +1,27 @@
 <?php
 
+declare(strict_types=1);
 
 namespace App\Model\User\UseCase\Role;
 
+use Symfony\Component\Validator\Constraints as Assert;
 
 class Command
 {
     /**
      * @var string
+     * @Assert\NotBlank()
      */
-    public $id;
+    public string $id;
 
     /**
      * @var string
+     * @Assert\NotBlank()
      */
-    public $role;
+    public string $role;
+
+    public function __construct(string $id)
+    {
+        $this->id = $id;
+    }
 }
