@@ -1,15 +1,15 @@
 <?php
 
+declare(strict_types=1);
 
-namespace App\Model\User\UseCase\SignUp\Confirm;
-
+namespace App\Model\User\UseCase\SignUp\Confirm\ByToken;
 
 class Command
 {
     /**
      * @var string
      */
-    public $token;
+    public string $token;
 
     public function __construct(string $token)
     {
