@@ -2,12 +2,11 @@
 
 declare(strict_types=1);
 
-namespace App\Model\User\UseCase\Role;
+namespace App\Model\User\UseCase\SignUp\Confirm\Manual;
 
-use App\Model\User\Entity\User\Id;
-use App\Model\User\Entity\User\Role;
-use App\Model\User\Entity\User\UserRepository;
 use App\Model\Flusher;
+use App\Model\User\Entity\User\Id;
+use App\Model\User\Entity\User\UserRepository;
 
 class Handler
 {
@@ -20,11 +19,11 @@ class Handler
         $this->flusher = $flusher;
     }
 
-    public function handle(Command $command): void
+    public function handle(Command $command)
     {
         $user = $this->users->get(new Id($command->id));
 
-        $user->changeRole(new Role($command->role));
+        $user->confirmSignUp();
 
         $this->flusher->flush();
     }
